+++ frontend/src/App.js
@@ -7,6 +7,7 @@ import Resources from "./pages/Resources";
 import Thumbnails from "./pages/Thumbnails";
 import Redeem from "./pages/Redeem";
 import Roadmap from "./pages/Roadmap";
+import Changelog from "./pages/Changelog";
 import SettingsKeys from "./pages/SettingsKeys";
 import Login from "./pages/Login";
 import Admin from "./pages/Admin";
@@ -194,6 +195,7 @@ export default function App() {
             <Route path="/resources" element={<RequireAuth><Resources /></RequireAuth>} />
             <Route path="/redeem" element={<Redeem />} />
             <Route path="/roadmap" element={<Roadmap />} />
+            <Route path="/changelog" element={<Changelog />} />
             <Route path="/settings/keys" element={<RequireAuth><SettingsKeys /></RequireAuth>} />
             <Route path="/admin" element={<RequireAuth><Admin /></RequireAuth>} />
             <Route path="/" element={<Navigate to="/scripts" replace />} />